Created preview and editor window + drop down menu fix.
Created a preview that gathers the data from the item that was clicked on and displays it in a solo window with all the information. Created a editor window which in the future can be sued to add to the listing. Fixed the drop down menu navigation to finally work (still not properly working on android) and added a shadow type of background when a drop menu is open.
Other bugs and fixes.